pyproject.toml
[project]
name = "scrape-articles"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "weasyprint>=63.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.17",
    "newspaper3k>=0.2.8",
    "python-docx>=1.1.2",
    "lxml-html-clean>=0.3.1",
    "jinja2>=3.1.4",
    "flask>=3.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
